Updated instinct limiter
Tenitively renamed to "neural inhibitor", though due to both the genericness of that name likely causing it to appear in other mods and the poor indication the name gives of it's true purpose, I think a better name is in order.  Quite proud of the description however.
Also, updated cost list with late-game parts (based intent on the thingSetMakerTags being slightly less rare than bionic parts) and made it spawn on removal for consistancy with other added parts.  Since there is no crafting recipe associated with the part yet, the cost list goes unused.