Crafting morphsilk + change hazsilk
Morphsilk takes 30 cloth, 1 slurry, and a lot of time. It's not a profitable recipie (you lose money).
Hazsilk now takes 30 cloth, 1 antigen, and a lot of time. Still not profitable.
Glowvine and antigen value increased a bit (8 vs 7, and 90 vs 80). Not a huge change given the difficulty.
Hazsilk is cheaper due to the new recipie (3 instead of 12.8)

The values need to be tested and probably balanced.